Further bullet-proofing.
Closes #2.

+++ django_odbc/turin/lxml_objectify_parse_ap_bulletin.py
@@ -42,6 +42,7 @@ def main():
         cursor.close()
     except OperationalError, err:
         # log this ... 
+        cursor = None
         print err
     
     '''
@@ -82,39 +83,40 @@ def main():
 #     connection.encoding = 'utf-8'
       # may not need this either ... 
 #     connection.stringformat = ODBC.NATIVE_UNICODE_STRINGFORMAT
-    cursor = connection.cursor()
-#     print 'Before:', cursor.rowcount
+    if cursor:
+        cursor = connection.cursor()
+#         print 'Before:', cursor.rowcount
     
-    # Check and see what the latest entry in DT is
-    cursor.execute('''SELECT top 1 * FROM dt_z_guide.apBulletin ORDER BY id DESC''')
-    [(latest_id, ap_timestamp, current_string, my_timestamp)] = cursor.fetchall()
-    current_string = current_string.encode('utf-8')
-    print u'Most recent in DT database:', latest_id, ap_timestamp, current_string, my_timestamp
+        # Check and see what the latest entry in DT is
+        cursor.execute('''SELECT top 1 * FROM dt_z_guide.apBulletin ORDER BY id DESC''')
+        [(latest_id, ap_timestamp, current_string, my_timestamp)] = cursor.fetchall()
+        current_string = current_string.encode('utf-8')
+        print u'Most recent in DT database:', latest_id, ap_timestamp, current_string, my_timestamp
     
 #     cursor.execute('''UPDATE dt_z_guide.apBulletin SET updateText = '%s', createdDateTime = '%s' WHERE ID=2''' % (last_graf, created))
 #     cursor.execute('''INSERT INTO dt_z_guide.apBulletin (updateText, createdDateTime) VALUES ('WASHINGTON — US to reopen 18 of 19 embassies, consulates shuttered this week due to terrorist threat.', '2013-08-09 17:45:52.59')''')
     
-    if (last_graf and current_string) and last_graf == current_string:
-        print 'No CHANGE'
-        # No change in update date, but we'll update logTimestamp
-        cursor.execute('''UPDATE dt_z_guide.apBulletin SET logTimestamp = '%s' WHERE id = %s''' % (datetime.now().strftime('%c'), latest_id))
-    else:
-        cursor.execute('''INSERT INTO dt_z_guide.apBulletin (updateText, createdDateTime, logTimestamp) VALUES ('%s', '%s', '%s')''' % (last_graf.decode('utf-8'), created, datetime.now().strftime('%c')))
-#     print 'After:', cursor.rowcount
+        if (last_graf and current_string) and last_graf == current_string:
+            print 'No CHANGE'
+            # No change in update date, but we'll update logTimestamp
+            cursor.execute('''UPDATE dt_z_guide.apBulletin SET logTimestamp = '%s' WHERE id = %s''' % (datetime.now().strftime('%c'), latest_id))
+        else:
+            cursor.execute('''INSERT INTO dt_z_guide.apBulletin (updateText, createdDateTime, logTimestamp) VALUES ('%s', '%s', '%s')''' % (last_graf.decode('utf-8'), created, datetime.now().strftime('%c')))
+    #     print 'After:', cursor.rowcount
     
-    connection.commit()
-    
-    # Clean up once a day at 3:15 a.m., 'cause, why not?
-    if (3, 15) == (time.localtime().tm_hour, time.localtime().tm_min):
-        cursor.execute('''SELECT TOP 25 Id FROM dt_z_guide.apBulletin ORDER BY Id DESC''')
-        nested_id_list = cursor.fetchall()
-        id_list = [story_id[0] for story_id in nested_id_list]
-        id_tuple = str(tuple(id_list))
-        cursor.execute('''DELETE FROM dt_z_guide.apBulletin WHERE id NOT IN %s''' % id_tuple)
-        print cursor.rowcount
         connection.commit()
     
-    cursor.close()
-    connection.close()
+        # Clean up once a day at 3:15 a.m., 'cause, why not?
+        if (3, 15) == (time.localtime().tm_hour, time.localtime().tm_min):
+            cursor.execute('''SELECT TOP 25 Id FROM dt_z_guide.apBulletin ORDER BY Id DESC''')
+            nested_id_list = cursor.fetchall()
+            id_list = [story_id[0] for story_id in nested_id_list]
+            id_tuple = str(tuple(id_list))
+            cursor.execute('''DELETE FROM dt_z_guide.apBulletin WHERE id NOT IN %s''' % id_tuple)
+            print cursor.rowcount
+            connection.commit()
+    
+        cursor.close()
+        connection.close()
 
 if __name__ == "__main__" : main()
